Calculate the fibonacci of 50000
By calculating the Fibonacci number of 50000 we can show that the function with Tail recursive has a better performance that the one which is a simple recursive function. Moreover the recursive functions throws an StackOverflow exception and the tail-recursive function doesn't.

This is due to the fact that the call in tail position (line 14) is compiled to an iterative loop automatically by Scala ans this iterative loop doesn't consume call stack frames for each iteration. On the other, the recursive function (specifically line 6) does consume call stack frames as it has to calculate the value of the return of the method plus the value of the other return of the method.